Adding unittest for `TasksTrigger.handle_specify_commands()` and refactoring unittest cases
